editing stylesheets

+++ sngconnect/appearance/views.py
@@ -8,6 +8,8 @@ from pyramid.view import view_config
 from sngconnect.translation import _
 from sngconnect.appearance import forms
 
+STYLESHEET_FILENAME = 'stylesheet.css'
+
 @view_config(
     route_name='sngconnect.appearance.appearance',
     renderer='sngconnect.appearance:templates/appearance.jinja2',
@@ -17,46 +19,82 @@ def appearance(request):
     assets_path = request.registry['settings'][
         'sngconnect.appearance_assets_upload_path'
     ]
+    stylesheet_file_path = os.path.join(assets_path, STYLESHEET_FILENAME)
+    stylesheet = None
+    try:
+        with open(stylesheet_file_path, 'r') as stylesheet_file:
+            stylesheet = stylesheet_file.read()
+    except IOError:
+        pass
+    update_stylesheet_form = forms.UpdateStylesheetForm(
+        stylesheet=stylesheet,
+        csrf_context=request
+    )
     upload_form = forms.UploadAssetForm(
         assets_path,
         csrf_context=request
     )
     if request.method == 'POST':
-        upload_form.process(request.POST)
-        if upload_form.validate():
-            filename = upload_form.get_filename()
-            input_file = upload_form.get_file()
-            output_file_path = os.path.join(assets_path, filename)
-            try:
-                os.makedirs(assets_path)
-            except OSError as exception:
-                if (exception.errno == errno.EEXIST and
-                        os.path.isdir(assets_path)):
-                    pass
-                else:
-                    raise
-            with open(output_file_path, 'wb') as output_file:
-                while True:
-                    data = input_file.read(2 << 16)
-                    if not data:
-                        break
-                    output_file.write(data)
-            request.session.flash(
-                _("New file has been succesfuly uploaded."),
-                queue='success'
-            )
-        else:
-            request.session.flash(
-                _(
-                    "There were some problems with your request."
-                    " Please check the form for error messages."
-                ),
-                queue='error'
-            )
+        if 'submit_update_stylesheet' in request.POST:
+            update_stylesheet_form.process(request.POST)
+            if update_stylesheet_form.validate():
+                with open(stylesheet_file_path, 'w') as output_file:
+                    output_file.write(update_stylesheet_form.stylesheet.data)
+                request.session.flash(
+                    _("Stylesheet has been succesfuly saved."),
+                    queue='success'
+                )
+                return httpexceptions.HTTPFound(
+                    request.route_url('sngconnect.appearance.appearance')
+                )
+            else:
+                request.session.flash(
+                    _(
+                        "There were some problems with your request."
+                        " Please check the form for error messages."
+                    ),
+                    queue='error'
+                )
+        elif 'submit_upload_asset' in request.POST:
+            upload_form.process(request.POST)
+            if upload_form.validate():
+                filename = upload_form.get_filename()
+                input_file = upload_form.get_file()
+                output_file_path = os.path.join(assets_path, filename)
+                try:
+                    os.makedirs(assets_path)
+                except OSError as exception:
+                    if (exception.errno == errno.EEXIST and
+                            os.path.isdir(assets_path)):
+                        pass
+                    else:
+                        raise
+                with open(output_file_path, 'wb') as output_file:
+                    while True:
+                        data = input_file.read(2 << 16)
+                        if not data:
+                            break
+                        output_file.write(data)
+                request.session.flash(
+                    _("New file has been succesfuly uploaded."),
+                    queue='success'
+                )
+                return httpexceptions.HTTPFound(
+                    request.route_url('sngconnect.appearance.appearance')
+                )
+            else:
+                request.session.flash(
+                    _(
+                        "There were some problems with your request."
+                        " Please check the form for error messages."
+                    ),
+                    queue='error'
+                )
     try:
-        filenames = os.listdir(assets_path)
+        filenames = set(os.listdir(assets_path))
     except OSError:
-        filenames = []
+        filenames = set()
+    filenames.discard(STYLESHEET_FILENAME)
     files = []
     for filename in filenames:
         file_path = os.path.join(assets_path, filename)
@@ -77,6 +115,7 @@ def appearance(request):
         })
     return {
         'files': files,
+        'update_stylesheet_form': update_stylesheet_form,
         'upload_form': upload_form,
     }
 
@@ -92,10 +131,11 @@ def delete_asset(request):
     delete_form = forms.DeleteAssetForm(csrf_context=request)
     delete_form.process(request.POST)
     if delete_form.validate():
-        try:
-            os.remove(os.path.join(assets_path, delete_form.filename.data))
-        except OSError:
-            pass
+        if delete_form.filename.data != STYLESHEET_FILENAME:
+            try:
+                os.remove(os.path.join(assets_path, delete_form.filename.data))
+            except OSError:
+                pass
         request.session.flash(
             _("File has been succesfuly deleted."),
             queue='success'
